Refactored DataLoader into two classes for unit testing vs live server usage
